Move classes to separate file
